Create `accounts_drones`
The `accounts_drones` table contains the drones of an account, this
way it's possible to keep track of a drone level.

This commit also improves the configuration items by identifying if an
item is equipped in a drone or a pet, if none of both is set, it's
assumed it's equipped in the ship. Also fixes relations of
`accounts_pets` -> `levels` and column values.